cleaned up StartGameButton and made Arrow svg reusable
