+++ src/components/Games/WhatIsThisGame.js
@@ -54,7 +54,7 @@ import {
 } from '../../const';
 
 import StarsToEarn from '../Stars/StarsToEarn';
-import StartGameButton from '../StartGameButton';
+import StartGameButton from '../StartGameButton/StartGameButton';
 import GameCompleteModal from '../GameCompleteModal/GameCompleteModal';
 import Loader from '../Loader/Loader';
